Synchronise media type bitfield with reality -- it's 64bits wide.
svn path=/trunk/libcss/; revision=6453

+++ include/libcss/stylesheet.h
@@ -19,7 +19,7 @@ typedef css_error (*css_import_handler)(void *pw, const char *url,
 
 css_error css_stylesheet_create(css_language_level level,
 		const char *charset, const char *url, const char *title,
-		css_origin origin, uint32_t media,
+		css_origin origin, uint64_t media,
 		css_import_handler import_callback, void *import_pw,
 		css_alloc alloc, void *alloc_pw, css_stylesheet **stylesheet);
 css_error css_stylesheet_destroy(css_stylesheet *sheet);
@@ -31,7 +31,7 @@ css_error css_stylesheet_data_done(css_stylesheet *sheet);
 css_error css_stylesheet_get_url(css_stylesheet *sheet, const char **url);
 css_error css_stylesheet_get_title(css_stylesheet *sheet, const char **title);
 css_error css_stylesheet_get_origin(css_stylesheet *sheet, css_origin *origin);
-css_error css_stylesheet_get_media(css_stylesheet *sheet, uint32_t *media);
+css_error css_stylesheet_get_media(css_stylesheet *sheet, uint64_t *media);
 
 css_error css_stylesheet_get_disabled(css_stylesheet *sheet, bool *disabled);
 css_error css_stylesheet_set_disabled(css_stylesheet *sheet, bool disabled);
